colli: RoboShape is now base on base-frame, not on laser anymore
The base-frame is "/base_link". It does not make sense to define the shape
around a "component" as the laser, but is reasonable to be defined around
the base of the robot.
Everyhing is based on that frame (search, obstacle size + position, actual
movement). This can also reduce some code where the colli tracks/passes both
laser-position and robot-position throughout its modules.

CAUTION:
As of now, EscapeMode will not work properly! It is purely based on laser-data.
Detection of collisions and decision about translation/rotation also relies
on a laser-based RoboShape. As the actual movement is around the base-frame though,
there is no way that this is safe, it will always require code-hacks and magic
numbers. We need the laser-data transformed to the base-frame in order to make
the EscapeMode work again (work at all, but also work better/properly).