openrave: use RefPtr instead of pointers
The OpenRave environment, robots and manipulators are created, modified
and shared among other threads. With pointers, we always needed to
remember deleting instances, but also be cautious NOT to delete when
we shouldn't (this is the more critical case).
Using RefPtr should simplify all this.

Other plugins might need to adapt!