jaco: delete all objects before cloning
For some reason, the second run of skill jaco_uncoord() resulted
in a null-pointer for a KinBody. There is probably a problem with
the environment-cloning, as the 2nd run deletes and adds new objects
in viewer_env -> probably faulty comparison and therefore not copied
to planner_env -> objects in planner_env have same name, but are not
"the same".
For now we just delete all objects in planner_env to avoid this. It
is fast enough, so there is not much of a downside here.

Additionally, when restoring robot-state, skip active manipulator
and DOF settings. We are mainly interested in link state and
grabbed bodies.