jaco: remove target-type TARGET_TRAJEC, every target can be a trajectory
Whether or not we want to plan a trajectory is now defined by the
trajec-state. This allows each target-type (angular, or cartesian, ...)
to be passed on to a planner.
We plan with euler-targets now, not with pure joint angles. This may
result in less errors, because now the planner can use multiple joint
configurations for the goal, not just the one we checked for IK earlier
(which was with less collision-avoidance btw).