jaco: each arm gets own full set of threads now
Up to now we had one act_thread and one info_thread handling all the
connected arms. That required many case considerations, and it was not
fully concurrent (for example checking incoming messages).
We now provide one of each threads for each connected arm. This gives us
better concurrency, and also reduces a lot of code.