jaco: add set of threads for coordinated bimanual manipulation
For uncoordinated manipulation, we needed separate set of threads
for each arm. We need another set though for the coordinated
manipulation (act_thread for interface-interaction and hooking,
openrave_thread for the planning, and goto_thread for the
coordinated movement execution).

Also renamed openrave-threads to
OpenraveThread (former OpenraveSingleThread) and
BimanualOpenraveThread (former OpenraveDualThread) to have the same
namings for threads.