openrave: add possibility to set a name for environment
This helps a lot for debugging, because now the name and the
unique environment-id are part of the output text (also when
throwing exceptions).
Otherwise it is hardly possible to tell which environment produced
which output, when having multiple environments.

Also added a few more debug messages and removed unnecessary brackets.