jaco: fix order of thread loading, avoiding loop time extensions
act_thread exceeded loop-time, because it ordered the openrave-threads
to update the robot model. However, this would take a while if the
environments were not ready yet. We now wait for the openrave-threads
to be initialized. This also requires that we load them in the correct
order, i.e. first the bimanual-thread (which loads the robot) and then
the threads for single-arm planning.