jaco: add additional paramter "followup" to set_joints() for JacoArm
This boolean parameter defines if the target is a single trajectory
point (i.e. the first of a trajectory), or a consecutive one.
If true, the driver needs to make sure that the API controls the arm
and that the mode is set to "angular". We must not do that for the
consecutive trajectory-points though, because then the FIFO queue would
not get filled fast enough -> arm stutters.