jaco: implement bimanual_goto_thread
This thread executes the movements that require coordination. Trajectories
are executed as soon as they are ready for both left and right arm. The
trajectory points are alternatingly passed to the arm, in order to make sure
that their internal queues are filled at the same rate and that there is no
delay.

As the targets are only removed from the queue when they are fully executed,
this also means that the final-field of the interfaces for single movement
also show "false" and "true" accordingly