webview-agent: Add the ability to save generated dot graphs
The webview-agent plugin can now save generated dot graphs, such that
the agent's behavior can be analyzed afterwards.

Image generation now happens in a separate worker thread, which is used
both the WebRequestProcessor and the plugin itself (which saves the
generated dot graphs). The default target directory for the dot graphs
is ./, all files have the naming pattern agent-$time.png, where $time is
the current time in msec.

The image generation takes a while and is in general slower than the
agent's execution. Therefore, a request queue is used, which reads the
values from the buffered interface. This means that the buffer size must
be chosen appropriately.