jaco: fix final-check of bimanual_goto_thread
We need to use the same checking we do for regular angular targets,
just as in goto_thread. That means, we need to make it TARGET_ANGULAR,
not leave it TARGET_GRIPPER. Simple reason: TARGET_GRIPPER relies on
the controller to say "final", which compares velocities, which is
not working properly yet (driver returns joint values instead of
velocities).