simple-obst-tracker: initial commit
The plugin simple obstacle tracker uses the laser-cluster plugin to calculate
the velocity of dynamic obstacles. A cfg-value specifies the number of
subsequent measurements that are used to calculate a moving average of the
velocity of a laser-cluster. The moving average is used as it behaves like
a low-pass filter.

For each Position3DInterface a corresponding Velocity3DInterface is opened.
The Velocity3DInterface holds:
- center of mass - x/y Position
- axial velocities (vx/vy)
- resulting absolute velocity

Only 2D movements are tracked.
Angular movements are not tracked.
There is no outlier correction at the moment.