obst-tracker: clear circular-buffer on visibility-history<val
Moving averager only calculates velocity if visibility-history of laser-cluster
is above a defined value (config).

If a cluster is lost - the visibility history drops. In this case, the moving-
average has to flush its internal circular-buffer. Otherwise it would use the
old position values if a new laser-cluster is detected.